# Conflicts:
#	package.json

+++ src/model_gen/modeling.ts
@@ -1,9 +1,9 @@
-import type { gp_Trsf, OpenCascadeInstance } from '$assets/replicad_single'
+import type { gp_Trsf, OpenCascadeInstance, TDataStd_Name } from '$assets/replicad_single'
 import { buildSewnSolid, makeTriangle } from '$lib/worker/modeling/index'
 import Trsf from '$lib/worker/modeling/transformation'
 import { stat } from 'fs/promises'
 import loadMF from 'manifold-3d'
-import type { CrossSection, Manifold, ManifoldToplevel } from 'manifold-3d'
+import type { Manifold, ManifoldToplevel } from 'manifold-3d'
 import {
   type AnyShape,
   cast,
@@ -24,7 +24,7 @@ import {
   Solid,
   Transformation,
 } from 'replicad'
-import { Matrix4, Triangle, Vector3 } from 'three'
+import { Matrix4, Vector3 } from 'three'
 
 let mf: ManifoldToplevel
 export async function loadManifold(): Promise<void> {
@@ -83,7 +83,7 @@ export function simplify(op: Operation) {
     case 'offset':
     case 'multmatrix':
     case 'translate':
-      op.obj = simplify(op.obj)
+      op.obj = simplify(op.obj)!
       if (op.obj == null) return null
       return op
   }
@@ -113,22 +113,24 @@ export function stringifyOperation(op: Operation, indent = 0): string {
   }
 }
 
-const filter = (x) => x.filter(s => s && s.length !== 0)
+type Op = Operation
+type N = number
+const filter = (x: Op[]) => x.filter(s => s && (s as any).length !== 0)
 
 // Build a syntax tree! Helps deal with duplicated nodes
-export const cube = (x, y, z) => ({ op: 'cube', x, y, z })
-export const cylinder = (r, h) => ({ op: 'cylinder', r, h })
-export const translate = (x, y, z, obj) => ({ op: 'translate', x, y, z, obj })
-export const mirror = (x, y, z, obj) => ({ op: 'mirror', x, y, z, obj })
-export const union = (shapes) => ({ op: 'union', shapes: filter(shapes) })
-export const difference = (shapes) => ({ op: 'difference', shapes: filter(shapes) })
-export const hull = (shapes) => ({ op: 'hull', shapes: filter(shapes) })
-export const square = (x, y) => ({ op: 'square', x, y })
-export const circle = (r) => ({ op: 'circle', r })
-export const extrudeLinear = (c, obj) => ({ op: 'linear_extrude', height: c.height, obj })
-export const extrudeRotate = (c, obj) => ({ op: 'rotate_extrude', angle: c.angle, obj })
-export const scale = (x, y, z, obj) => ({ op: 'multmatrix', matrix: to2d(new Matrix4().makeScale(x, y, z).transpose().elements), obj })
-export const rotate = (a, x, y, z, obj) => ({ op: 'multmatrix', matrix: to2d(new Trsf().rotate(a * 180 / Math.PI, [0, 0, 0], [x, y, z]).matrix()), obj })
+export const cube = (x: N, y: N, z: N) => ({ op: 'cube', x, y, z })
+export const cylinder = (r: N, h: N) => ({ op: 'cylinder', r, h })
+export const translate = (x: N, y: N, z: N, obj: Op[]) => ({ op: 'translate', x, y, z, obj })
+export const mirror = (x: N, y: N, z: N, obj: Op[]) => ({ op: 'mirror', x, y, z, obj })
+export const union = (shapes: Op[]) => ({ op: 'union', shapes: filter(shapes) })
+export const difference = (shapes: Op[]) => ({ op: 'difference', shapes: filter(shapes) })
+export const hull = (shapes: Op[]) => ({ op: 'hull', shapes: filter(shapes) })
+export const square = (x: N, y: N) => ({ op: 'square', x, y })
+export const circle = (r: N) => ({ op: 'circle', r })
+export const extrudeLinear = (c: any, obj: Op[]) => ({ op: 'linear_extrude', height: c.height, obj })
+export const extrudeRotate = (c: any, obj: Op[]) => ({ op: 'rotate_extrude', angle: c.angle, obj })
+export const scale = (x: N, y: N, z: N, obj: Op[]) => ({ op: 'multmatrix', matrix: to2d(new Matrix4().makeScale(x, y, z).transpose().elements), obj })
+export const rotate = (a: N, x: N, y: N, z: N, obj: Op[]) => ({ op: 'multmatrix', matrix: to2d(new Trsf().rotate(a * 180 / Math.PI, [0, 0, 0], [x, y, z]).matrix()), obj })
 
 const to2d = (m: number[]) => [m.slice(0, 4), m.slice(4, 8), m.slice(8, 12), m.slice(12, 16)]
 
@@ -269,7 +271,7 @@ export function serialize(filename: string, model: AnyShape) {
   writer.Model(true).delete()
   const progress = new oc.Message_ProgressRange_1()
 
-  writer.Transfer(model.wrapped, oc.STEPControl_StepModelType.STEPControl_AsIs, true, progress)
+  writer.Transfer(model.wrapped, oc.STEPControl_StepModelType.STEPControl_AsIs as any, true, progress)
 
   // Convert to a .STEP File
   const done = writer.Write(filename + '.step')
@@ -316,27 +318,38 @@ export async function importSTEPSpecifically(blob: Blob, name: string) {
   const oc = getOC() as OpenCascadeInstance
   const [r, gc] = localGC()
 
-  const text = await blob.text()
-  const re = new RegExp(`(#\\d*)\\s*=\\s*NEXT_ASSEMBLY_USAGE_OCCURRENCE\\([^\n;]*,'${name}'`)
-  const label = text.match(re)![1]
+  // const text = await blob.text()
+  // // const re = new RegExp(`(#\\d*)\\s*=\\s*NEXT_ASSEMBLY_USAGE_OCCURRENCE\\([^\n;]*,'${name}'`)
+  // const re = new RegExp(`(#\\d*)\\s*=\\s*PRODUCT\\('${name}'`)
+  // const label = text.match(re)![1]
 
   const fileName = Date.now().toString(36) + Math.random().toString(36).substring(2)
   const bufferView = new Uint8Array(await blob.arrayBuffer())
   oc.FS.writeFile(`/${fileName}`, bufferView)
 
-  const reader = r(new oc.STEPControl_Reader_1())
-  if (reader.ReadFile(fileName)) {
+  const reader = r(new oc.STEPCAFControl_Reader_1())
+  const doc = r(new oc.TDocStd_Document(new oc.TCollection_ExtendedString_2('XmlOcaf', false)))
+  if (reader.ReadFile(fileName) && reader.Transfer_1(new oc.Handle_TDocStd_Document_2(doc), new oc.Message_ProgressRange_1())) {
     oc.FS.unlink('/' + fileName)
 
-    const model = reader.StepModel().get()
-    const rank = model.NextNumberForLabel(label, 0, false)
-    if (rank == 0) throw new Error(`Model ${label} not found`)
-    reader.TransferOne(rank, r(new oc.Message_ProgressRange_1()))
-    const stepShape = r(reader.OneShape())
+    const shapeTool = oc.XCAFDoc_DocumentTool.ShapeTool(doc.Main()).get()
+    const labels = r(new oc.TDF_LabelSequence_1())
+    shapeTool.GetShapes(labels)
+    for (let i = 1; i <= labels.Length(); i++) {
+      const label = labels.Value(i)
+      const handle = r(new oc.Handle_TDF_Attribute_1())
+      label.FindAttribute_1(oc.TDataStd_Name.GetID(), handle)
+      const nameObj = handle.get() as TDataStd_Name
+      const nameStr = r(new oc.TCollection_AsciiString_13(nameObj.Get(), 0)).ToCString()
+      if (nameStr == name) {
+        const shape = cast(r(oc.XCAFDoc_ShapeTool.GetShape_2(label)))
+        gc()
+        return shape
+      }
+    }
 
-    const shape = cast(stepShape)
     gc()
-    return shape
+    throw new Error(`Model "${name}" not found`)
   } else {
     oc.FS.unlink('/' + fileName)
     gc()
